requirements.txt
python-telegram-bot==21.0.1
requests==2.31.0